% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellBayes.R
\name{constraintFn}
\alias{constraintFn}
\title{This defines the equality constraint for the vector of proportion coefficients}
\usage{
constraintFn(p)
}
\arguments{
\item{p}{proportion coeefficiento of DMM}
}
\value{
summation
}
\description{
This defines the equality constraint for the vector of proportion coefficients
}
