% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellBoost.R
\name{detectFlagCells}
\alias{detectFlagCells}
\title{TIMING SPEED UP A LITTLE BIT??
The function detects the droplet barcode that are flag Cells marked by its expression of flag genes>median+3sd}
\usage{
detectFlagCells(rawDGE, mitoGenes = TRUE, otherFlagGenes = NULL)
}
\arguments{
\item{rawDGE}{the originalDGE that returned from prepareData function}

\item{mitoGenes}{If TRUE, mitochondrial genes will be flagged genes, regardless of mouse mt or human mt.}

\item{otherFlagGenes}{specify the genes that needs to be flagged by user. the function automatically detect flagcells by mitochondrial genes. This parameter gives
extra flaggens that characterize unwanted cells.}
}
\value{
this function returns a vector with barcodes of flag cells
}
\description{
TIMING SPEED UP A LITTLE BIT??
The function detects the droplet barcode that are flag Cells marked by its expression of flag genes>median+3sd
}
