% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellBoost.R
\name{detectOutlier}
\alias{detectOutlier}
\title{This function detect barcodes with pseudo positive labels}
\usage{
detectOutlier(rawDGE, rawCounts, expectedN = 1000)
}
\arguments{
\item{rawDGE}{digital expression matrix of the original data}

\item{rawCounts}{a digital expression matrix with whole data points}

\item{expectedN}{number of expected cell-containing droplets}
}
\value{
returns xxx
}
\description{
This function detect barcodes with pseudo positive labels
}
