% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dgeShuffle}
\alias{dgeShuffle}
\title{This function generate shuffled DGE}
\usage{
dgeShuffle(n_genes, n_bcds, num_nonzeros, numis, geneInd, barcodeInd, totalumi)
}
\arguments{
\item{n_genes}{number of genes}

\item{n_bcds}{number of barcodes}

\item{num_nonzeros}{number of non-zero umis}

\item{numis}{vector of umis}

\item{geneInd}{vector of gene index}

\item{barcodeInd}{index of barcoders}

\item{totalumi}{sum of umis}
}
\value{
a list
}
\description{
This function generate shuffled DGE
}
