% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getAmbientProp}
\alias{getAmbientProp}
\title{Rcpp function to get the gene profile of ambient droplets}
\usage{
getAmbientProp(m, nrow, ncol)
}
\arguments{
\item{m}{a sparse matrix of ambient drpolets}

\item{nrow}{the rows of m}

\item{ncol}{the cols of m}
}
\description{
Rcpp function to get the gene profile of ambient droplets
}
