% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellBayes.R
\name{getP}
\alias{getP}
\title{This function get the proportion coefficient for Multinomial Mixture Model}
\usage{
getP(geneProfile, mtx, threshold = 100, seed = 0, ub_p = 1, lb_p = 0)
}
\arguments{
\item{geneProfile}{matrix of gene profile for all cell types and soup}

\item{mtx}{DGE matrix}

\item{threshold}{focus on droplets with UMI>thershold}

\item{seed}{seed for generating starting point}

\item{ub_p}{a scalar of upper bound of p}

\item{lb_p}{a scalar of lower bound of p}
}
\value{
matrix of estimated proportion of DMM
}
\description{
This function get the proportion coefficient for Multinomial Mixture Model
}
