% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellBoost.R
\name{labelData}
\alias{labelData}
\title{This function generate positive and negative pseudo-labels.}
\usage{
labelData(
  rawDGE,
  shfDGE,
  rawPC,
  shfPC,
  HVG,
  flagCells,
  outlier,
  dataName,
  expectedN
)
}
\arguments{
\item{rawDGE}{digital expression matrix of the original data}

\item{shfDGE}{digital expression matrix of the shuffle data}

\item{rawPC}{top one 100PCs of raw DGE}

\item{shfPC}{top one 100PCs of shuffle DGE}

\item{HVG}{a vector with highly variable genes}

\item{flagCells}{a vector of flagged cells}

\item{outlier}{a vector of barcodes that are identified as "pseudo-positive"}

\item{dataName}{data name}

\item{expectedN}{expected number of cell-containing droplets}
}
\value{
returns xxx
}
\description{
This function generate positive and negative pseudo-labels.
}
