% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellBoost.R
\name{logNormalizeUMI}
\alias{logNormalizeUMI}
\title{The function log normalize the DGE with only Highly variable genes}
\usage{
logNormalizeUMI(UMIcount, HVG)
}
\arguments{
\item{UMIcount}{a DGE}

\item{HVG}{vector of highly variable genes}
}
\value{
the functions returns a log normalized DGE
}
\description{
The function log normalize the DGE with only Highly variable genes
}
