% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellBayes.R, R/SiftCellMix.R
\name{objectiveFn}
\alias{objectiveFn}
\title{This function defines the objective function}
\usage{
objectiveFn(p, umi, geneProfile)

objectiveFn(p, umi, geneProfile)
}
\arguments{
\item{p}{proportion coeefficiento of DMM}

\item{umi}{a vector of UMIs}

\item{geneProfile}{this is a transposed geneProfile(as a matrix)}
}
\value{
negative log likelihood of DMM

negative log likelihood of DMM
}
\description{
This function defines the objective function

This function defines the objective function
}
