% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellBoost.R
\name{prepData}
\alias{prepData}
\title{The function prepares data for Original and shuffled DGE}
\usage{
prepData(workingdir, UMIthreshold = 100)
}
\arguments{
\item{workingdir}{the working directory with two folders named DGE and shuffleDGE. The shuffleDGE canbe
generated byxxx}

\item{UMIthreshold}{a UMI cutoff that filter out droplet with total UMI below the cutoff}
}
\value{
the function returns a list of matrix:originalDGE and shuffleDGE of droplets with UMI bigger or
equal than the UMI threshold. In addition a matrix with all droplets is also returned.
}
\description{
The function prepares data for Original and shuffled DGE
}
