% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runDMM}
\alias{runDMM}
\title{Rcpp function to get the coefficieint of DMM}
\usage{
runDMM(geneProfile, mtx, ub_p, lb_p)
}
\arguments{
\item{geneProfile}{matrix of gene profile for all cell types and soup}

\item{mtx}{a DGE matrix}

\item{ub_p}{a vector of upper bound for p}

\item{lb_p}{a vector of lower bound for p}
}
\value{
a matrix with convergence and coefficients
}
\description{
Rcpp function to get the coefficieint of DMM
}
