% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellBoost.R
\name{runPCA}
\alias{runPCA}
\title{The function get the top 100 PC of HVG for both original and shuffled DGE}
\usage{
runPCA(
  rawDGE,
  shfDGE,
  HVG,
  log = TRUE,
  dataName,
  retx = TRUE,
  center = TRUE,
  scale = TRUE,
  N.PCs = 100
)
}
\arguments{
\item{rawDGE}{original DGE}

\item{shfDGE}{shuffled DGE}

\item{HVG}{dataframe with gene name, tarones statistic(out)}

\item{log}{log Normalized the data. This is true by default.}

\item{dataName}{the dataName of the dataset. For example:PBMC, Brain}

\item{retx}{TRUE}

\item{center}{whether to center the data}

\item{scale}{whther to scale the data}

\item{N.PCs}{number of PCs}
}
\value{
the funtion returns PCs of rawDGE and shfDGE
}
\description{
The function get the top 100 PC of HVG for both original and shuffled DGE
}
