% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellBoost.R
\name{runXGBoostPy}
\alias{runXGBoostPy}
\title{This function source python code and run XGBoost}
\usage{
runXGBoostPy(train_path, test_path, workingdir, dataName)
}
\arguments{
\item{train_path}{path of training data}

\item{test_path}{path of test data}

\item{workingdir}{working directory}

\item{dataName}{name of the project}
}
\value{
a vector of cell containing droplets
}
\description{
This function source python code and run XGBoost
}
