% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellBayes.R
\name{SiftCellBayes}
\alias{SiftCellBayes}
\title{This function runs SiftCellBayes}
\usage{
SiftCellBayes(
  workingdir,
  celltype,
  alpha = 0.01,
  threshold = 100,
  ub_p = 1,
  lb_p = 0
)
}
\arguments{
\item{workingdir}{directory with DGE files}

\item{celltype}{dataframe with two columns, 1st column is cell barcode, 2nd column is cell type}

\item{alpha}{scalar between 0 and 1}

\item{threshold}{focus on droplets with UMI>threshold}

\item{ub_p}{upper bound of proportion. Default is 1}

\item{lb_p}{lower bound of proportion. Default is 0}
}
\value{
estimated proportion
}
\description{
This function runs SiftCellBayes
}
