% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellBoost.R
\name{SiftCellBoost}
\alias{SiftCellBoost}
\title{This function runs SiftCell-Boost, it generates a file of barcode of infered cell-containing droplets}
\usage{
SiftCellBoost(
  workingdir,
  threshold = 100,
  expectedN = 1000,
  dataName = "Sample",
  mitoGenes = TRUE,
  otherFlagGenes = NULL,
  N.PCs = 100,
  geneUMIthreshold = 50,
  seed = 0,
  quantileReg = FALSE
)
}
\arguments{
\item{workingdir}{working director with DGE and shuffleDGE two folders. Within DGE or shuffleDGE foler,}

\item{threshold}{UMI threshold; By default threshold = 100}

\item{expectedN}{number of expected cell-containing droplets in the dataset.}

\item{dataName}{name of the dataset. For example:PBMC}

\item{mitoGenes}{Default true}

\item{otherFlagGenes}{f0lag genes that characterize unwanted celltypes. For example:PPBP, a marker}

\item{N.PCs}{PCs}

\item{geneUMIthreshold}{genes less than the threshold is discarded}

\item{seed}{seed for randomization}

\item{quantileReg}{default if use top 1000 HVG}
}
\description{
This function runs SiftCell-Boost, it generates a file of barcode of infered cell-containing droplets
}
