% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellShuffle.R
\name{SiftCellShuffle}
\alias{SiftCellShuffle}
\title{The function generate shuffleDGE}
\usage{
SiftCellShuffle(workingdir)
}
\arguments{
\item{workingdir}{the working directory}
}
\description{
The function generate shuffleDGE
}
