% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellMix.R
\name{solveDMM}
\alias{solveDMM}
\title{The function solves the p by solnp in Rsolnp package}
\usage{
solveDMM(p0, umi, geneProfile, lb_p, ub_p)
}
\arguments{
\item{p0}{The startin points}

\item{umi}{vector of UMIs in a droplet}

\item{geneProfile}{matrix of gene profile for all cell types and soup}

\item{lb_p}{vector of lower bound}

\item{ub_p}{vecetor of upper bound}
}
\value{
vector of estimated proportion and convergence value
}
\description{
The function solves the p by solnp in Rsolnp package
}
