% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellMix.R
\name{objectiveFn}
\alias{objectiveFn}
\title{This function defines the objective function}
\usage{
objectiveFn(p, umi, geneProfile)
}
\arguments{
\item{p}{proportion coeefficient of Multinomial Mixture (MM) model}

\item{umi}{a vector of UMIs}

\item{geneProfile}{this is a transposed geneProfile(as a matrix)}
}
\value{
negative log likelihood of MM
}
\description{
This function defines the objective function
}
