% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiftCellBoost.R
\name{readDGE}
\alias{readDGE}
\title{This function read the DGE}
\usage{
readDGE(datadir = NULL)
}
\arguments{
\item{datadir}{the directory that includes the digital expression matrix. Version2 & Version 3}
}
\value{
a matrix
}
\description{
This function read the DGE
}
